+++ examples/private_zonal_with_networking/main.tf
@@ -14,9 +14,18 @@
  * limitations under the License.
  */
 
+data "google_client_config" "default" {}
+
+provider "kubernetes" {
+  load_config_file       = false
+  host                   = "https://${module.gke.endpoint}"
+  token                  = data.google_client_config.default.access_token
+  cluster_ca_certificate = base64decode(module.gke.ca_certificate)
+}
+
 module "gcp-network" {
   source       = "terraform-google-modules/network/google"
-  version      = "~> 2.0"
+  version      = "~> 3.1"
   project_id   = var.project_id
   network_name = var.network
 
@@ -51,7 +60,7 @@ data "google_compute_subnetwork" "subnetwork" {
 }
 
 module "gke" {
-  source     = "../../modules/beta-private-cluster/"
+  source     = "../../modules/private-cluster/"
   project_id = var.project_id
   name       = var.cluster_name
   regional   = false
@@ -73,7 +82,4 @@ module "gke" {
       display_name = "VPC"
     },
   ]
-}
-
-data "google_client_config" "default" {
 }