+++ examples/simple_regional_private_beta/main.tf
@@ -19,13 +19,22 @@ locals {
 }
 
 provider "google" {
-  version = "~> 3.35.0"
+  version = "~> 3.42.0"
   region  = var.region
 }
 
 provider "google-beta" {
-  version = "~> 3.35.0"
+  version = "~> 3.49.0"
   region  = var.region
+}
+
+data "google_client_config" "default" {}
+
+provider "kubernetes" {
+  load_config_file       = false
+  host                   = "https://${module.gke.endpoint}"
+  token                  = data.google_client_config.default.access_token
+  cluster_ca_certificate = base64decode(module.gke.ca_certificate)
 }
 
 data "google_compute_subnetwork" "subnetwork" {
@@ -60,7 +69,4 @@ module "gke" {
   cloudrun          = var.cloudrun
   dns_cache         = var.dns_cache
   gce_pd_csi_driver = var.gce_pd_csi_driver
-}
-
-data "google_client_config" "default" {
 }