+++ modules/asm/variables.tf
@@ -34,40 +34,10 @@ variable "location" {
   type        = string
 }
 
-variable "skip_gcloud_download" {
-  description = "Whether to skip downloading gcloud (assumes gcloud and kubectl already available outside the module)"
-  type        = bool
-  default     = true
-}
-
 variable "gcloud_sdk_version" {
   description = "The gcloud sdk version to use. Minimum required version is 293.0.0"
   type        = string
   default     = "296.0.1"
-}
-
-variable "enable_gke_hub_registration" {
-  description = "Enables GKE Hub Registration when set to true"
-  type        = bool
-  default     = true
-}
-
-variable "gke_hub_sa_name" {
-  description = "Name for the GKE Hub SA stored as a secret `creds-gcp` in the `gke-connect` namespace."
-  type        = string
-  default     = "gke-hub-sa"
-}
-
-variable "gke_hub_membership_name" {
-  description = "Memebership name that uniquely represents the cluster being registered on the Hub"
-  type        = string
-  default     = "gke-asm-membership"
-}
-
-variable "internal_ip" {
-  description = "Use internal ip for the cluster endpoint."
-  type        = bool
-  default     = false
 }
 
 variable "asm_dir" {
@@ -76,8 +46,19 @@ variable "asm_dir" {
   default     = "asm-dir"
 }
 
+variable "service_account_key_file" {
+  description = "Path to service account key file to auth as for running `gcloud container clusters get-credentials`."
+  default     = ""
+}
+
 variable "asm_version" {
   description = "ASM version to deploy. Available versions are documented in https://github.com/GoogleCloudPlatform/anthos-service-mesh-packages"
   type        = string
-  default     = "release-1.6-asm"
+  default     = "1.8"
+}
+
+variable "managed" {
+  description = "Whether the control plane should be managed."
+  type        = bool
+  default     = false
 }