+++ modules/k8s-operator-crd-support/main.tf
@@ -15,22 +15,22 @@
  */
 
 locals {
-  cluster_endpoint              = "https://${var.cluster_endpoint}"
-  private_key                   = var.create_ssh_key && var.ssh_auth_key == null ? tls_private_key.k8sop_creds[0].private_key_pem : var.ssh_auth_key
-  k8sop_creds_secret_key        = var.secret_type == "cookiefile" ? "cookie_file" : var.secret_type
-  should_download_manifest      = var.operator_path == null ? true : false
-  manifest_path                 = local.should_download_manifest ? "${path.root}/.terraform/tmp/${var.project_id}-${var.cluster_name}/config-management-operator.yaml" : var.operator_path
-  sync_branch_node              = var.sync_branch != "" ? format("syncBranch: %s", var.sync_branch) : ""
-  policy_dir_node               = var.policy_dir != "" ? format("policyDir: %s", var.policy_dir) : ""
-  hierarchy_controller_map_node = var.hierarchy_controller == null ? "" : format("hierarchy_controller:\n    %s", yamlencode(var.hierarchy_controller))
-  source_format_node            = var.source_format != "" ? format("sourceFormat: %s", var.source_format) : ""
+  cluster_endpoint                               = "https://${var.cluster_endpoint}"
+  private_key                                    = var.create_ssh_key && var.ssh_auth_key == null ? tls_private_key.k8sop_creds[0].private_key_pem : var.ssh_auth_key
+  k8sop_creds_secret_key                         = var.secret_type == "cookiefile" ? "cookie_file" : var.secret_type
+  should_download_manifest                       = var.operator_path == null ? true : false
+  manifest_path                                  = local.should_download_manifest ? "${path.root}/.terraform/tmp/${var.project_id}-${var.cluster_name}/config-management-operator.yaml" : var.operator_path
+  sync_branch_node                               = var.sync_branch != "" ? format("syncBranch: %s", var.sync_branch) : ""
+  policy_dir_node                                = var.policy_dir != "" ? format("policyDir: %s", var.policy_dir) : ""
+  hierarchy_controller_map_node                  = var.hierarchy_controller == null ? "" : format("hierarchyController:\n    %s", indent(4, replace(yamlencode(var.hierarchy_controller), "/((?:^|\n)[\\s-]*)\"([\\w-]+)\":/", "$1$2:")))
+  source_format_node                             = var.source_format != "" ? format("sourceFormat: %s", var.source_format) : ""
+  append_arg_use_existing_context_for_gatekeeper = var.use_existing_context ? "USE_EXISTING_CONTEXT_ARG" : ""
 }
 
 module "k8sop_manifest" {
-  source        = "terraform-google-modules/gcloud/google"
-  version       = "~> 1.3"
-  enabled       = local.should_download_manifest
-  skip_download = var.skip_gcloud_download
+  source  = "terraform-google-modules/gcloud/google"
+  version = "~> 2.0.2"
+  enabled = local.should_download_manifest
 
   create_cmd_entrypoint  = "gsutil"
   create_cmd_body        = "cp ${var.operator_latest_manifest_url} ${local.manifest_path}"
@@ -40,13 +40,14 @@ module "k8sop_manifest" {
 
 
 module "k8s_operator" {
-  source            = "terraform-google-modules/gcloud/google//modules/kubectl-wrapper"
-  version           = "~> 1.4"
-  module_depends_on = [module.k8sop_manifest.wait, var.cluster_endpoint]
-  skip_download     = var.skip_gcloud_download
-  cluster_name      = var.cluster_name
-  cluster_location  = var.location
-  project_id        = var.project_id
+  source                   = "terraform-google-modules/gcloud/google//modules/kubectl-wrapper"
+  version                  = "~> 2.0.2"
+  module_depends_on        = [module.k8sop_manifest.wait, var.cluster_endpoint]
+  cluster_name             = var.cluster_name
+  cluster_location         = var.location
+  project_id               = var.project_id
+  service_account_key_file = var.service_account_key_file
+  use_existing_context     = var.use_existing_context
 
   kubectl_create_command  = "kubectl apply -f ${local.manifest_path}"
   kubectl_destroy_command = "kubectl delete -f ${local.manifest_path}"
@@ -60,15 +61,18 @@ resource "tls_private_key" "k8sop_creds" {
 }
 
 module "k8sop_creds_secret" {
-  source            = "terraform-google-modules/gcloud/google//modules/kubectl-wrapper"
-  version           = "~> 1.4"
-  module_depends_on = [module.k8s_operator.wait]
-  skip_download     = var.skip_gcloud_download
-  cluster_name      = var.cluster_name
-  cluster_location  = var.location
-  project_id        = var.project_id
+  source  = "terraform-google-modules/gcloud/google//modules/kubectl-wrapper"
+  version = "~> 2.0.2"
 
-  kubectl_create_command  = "kubectl create secret generic ${var.operator_credential_name} -n=${var.operator_credential_namespace} --from-literal=${local.k8sop_creds_secret_key}='${local.private_key}'"
+  enabled                  = var.create_ssh_key == true || var.ssh_auth_key != null ? "true" : "false"
+  module_depends_on        = [module.k8s_operator.wait]
+  cluster_name             = var.cluster_name
+  cluster_location         = var.location
+  project_id               = var.project_id
+  service_account_key_file = var.service_account_key_file
+  use_existing_context     = var.use_existing_context
+
+  kubectl_create_command  = local.private_key != null ? "kubectl create secret generic ${var.operator_credential_name} -n=${var.operator_credential_namespace} --from-literal=${local.k8sop_creds_secret_key}='${local.private_key}'" : ""
   kubectl_destroy_command = "kubectl delete secret ${var.operator_credential_name} -n=${var.operator_credential_namespace}"
 }
 
@@ -90,21 +94,33 @@ data "template_file" "k8sop_config" {
   }
 }
 
-resource "local_file" "operator_cr" {
-  content  = data.template_file.k8sop_config.rendered
-  filename = "${path.module}/${var.project_id}-${var.cluster_name}/operator_cr.yaml"
+module "k8sop_config" {
+  source                   = "terraform-google-modules/gcloud/google//modules/kubectl-wrapper"
+  version                  = "~> 2.0.2"
+  module_depends_on        = [module.k8s_operator.wait, module.k8sop_creds_secret.wait]
+  cluster_name             = var.cluster_name
+  cluster_location         = var.location
+  project_id               = var.project_id
+  create_cmd_triggers      = { configmanagement = data.template_file.k8sop_config.rendered }
+  service_account_key_file = var.service_account_key_file
+  use_existing_context     = var.use_existing_context
+
+  kubectl_create_command  = "kubectl apply -f - <<EOF\n${data.template_file.k8sop_config.rendered}EOF"
+  kubectl_destroy_command = "kubectl delete -f - <<EOF\n${data.template_file.k8sop_config.rendered}EOF"
 }
 
-module "k8sop_config" {
-  source              = "terraform-google-modules/gcloud/google//modules/kubectl-wrapper"
-  version             = "~> 1.4"
-  module_depends_on   = [module.k8s_operator.wait, module.k8sop_creds_secret.wait]
-  skip_download       = var.skip_gcloud_download
-  cluster_name        = var.cluster_name
-  cluster_location    = var.location
-  project_id          = var.project_id
-  create_cmd_triggers = { configmanagement = local_file.operator_cr.content }
+module "wait_for_gatekeeper" {
+  source                   = "terraform-google-modules/gcloud/google//modules/kubectl-wrapper"
+  version                  = "~> 2.0.2"
+  enabled                  = var.enable_policy_controller ? true : false
+  module_depends_on        = [module.k8sop_config.wait]
+  cluster_name             = var.cluster_name
+  cluster_location         = var.location
+  project_id               = var.project_id
+  create_cmd_triggers      = { script_sha1 = sha1(file("${path.module}/scripts/wait_for_gatekeeper.sh")) }
+  service_account_key_file = var.service_account_key_file
+  use_existing_context     = var.use_existing_context
 
-  kubectl_create_command  = "kubectl apply -f ${local_file.operator_cr.filename}"
-  kubectl_destroy_command = "kubectl delete -f ${local_file.operator_cr.filename}"
+  kubectl_create_command  = "${path.module}/scripts/wait_for_gatekeeper.sh ${var.project_id} ${var.cluster_name} ${var.location} ${local.append_arg_use_existing_context_for_gatekeeper}"
+  kubectl_destroy_command = ""
 }