+++ modules/workload-identity/variables.tf
@@ -15,7 +15,7 @@
  */
 
 variable "name" {
-  description = "Name for both service accounts"
+  description = "Name for both service accounts. The GCP SA will be truncated to the first 30 chars if necessary."
   type        = string
 }
 
@@ -54,8 +54,20 @@ variable "use_existing_k8s_sa" {
   type        = bool
 }
 
+variable "annotate_k8s_sa" {
+  description = "Annotate the kubernetes service account with 'iam.gke.io/gcp-service-account' annotation. Valid in cases when an existing SA is used."
+  default     = true
+  type        = bool
+}
+
 variable "automount_service_account_token" {
   description = "Enable automatic mounting of the service account token"
   default     = false
   type        = bool
+}
+
+variable "roles" {
+  type        = list(string)
+  default     = []
+  description = "(optional) A list of roles to be added to the created Service account"
 }